+++ repo_health/__init__.py
@@ -1,22 +1,19 @@
 # pylint: disable=django-not-configured
+
 """
 This package contains checks for edx repo standards
 """
+
 import codecs
 import glob
 import os
+
 from configparser import ConfigParser
 
 import dockerfile
 import pytest
 
 __version__ = "1.0.0"
-
-
-GITHUB_URL_PATTERN = r"github.com[/:](?P<org_name>[^/]+)/(?P<repo_name>[^/]+).*#egg=(?P<package>[^\/]+).*"
-PYPI_PACKAGE_PATTERN = r"(?P<package_name>[^\/]+)==(?P<version>[^\/]+)"
-DJANGO_DEPS_SHEET_URL = "https://docs.google.com/spreadsheets/d/" \
-                        "19-BzpcX3XvqlazHcLhn1ZifBMVNund15EwY3QQM390M/export?format=csv"
 
 
 def parse_config_file(path):