+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 +++++
 
 * Added a check for the url= and project_urls= settings in setup.py and setup.cfg.
+* Read yaml files for both orgs while creating dashboard
 
 * Added a check for repos that indicate inclusion in Open edX in their openedx.yaml file, but aren't in the openedx GitHub organization.